[ALPS03305201] cam: fix KASAN
[Detail] fix KASAN slab-out-of-bounds

[Solution] done

MTK-Commit-Id: d81020fa4c829732bb62ecfad6de66df80c8860d

CR-Id: ALPS03305201
Feature: Others

+++ drivers/misc/mediatek/imgsensor/src/mt6763/imx362_mipi_raw/imx362mipi_Sensor.c
@@ -2119,9 +2119,9 @@ static kal_uint32 feature_control(MSDK_SENSOR_FEATURE_ENUM feature_id,
 		*feature_para_len=4;
 		break;
 	case SENSOR_FEATURE_SET_FRAMERATE:
-		LOG_INF("current fps :%d\n",  (UINT32)*feature_data);
+		LOG_INF("current fps :%llu\n", *feature_data);
 		spin_lock(&imgsensor_drv_lock);
-		imgsensor.current_fps = *feature_data;
+		imgsensor.current_fps = (UINT16)*feature_data;
 		spin_unlock(&imgsensor_drv_lock);
 		break;
 	case SENSOR_FEATURE_SET_HDR:
